requirements.txt
# Core dependencies for background remover AI demo
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
timm>=0.9.0
huggingface_hub>=0.15.0

# FastAPI and web dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Optional: CUDA support (uncomment if using GPU)
# torch-audio>=2.0.0

# Additional useful packages
numpy>=1.21.0
python-multipart